Fix incremental build
I noticed that VS was doing a full rebuild each time it runs the app.
There are a couple things needed to fix this:

Tweak the project file to not copy onlinefonts.js always. Only copy when
it has changed.

Change the Xaml files to not copy to output (they're not needed for
runtime and get compiled to a binary format).